VFIO: accept 64-bit CORB/RIRB addresses written as two words
This resolves an issue where all data was bogus with a recent Windows10
VM. This would write the upper and then lower address registers as two
32-bit writes.

I don't know if anyone writes both at once with a 64-bit write but this
should handle it.

Variables renamed to match new usage.